ID-> Id DBService improvements

+++ src/main/java/com/evaluation/service/db/model/IpBlacklist.java
@@ -1,7 +1,5 @@
 package com.evaluation.service.db.model;
 
-import java.math.BigInteger;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -12,13 +10,13 @@ import javax.persistence.Table;
 public class IpBlacklist {
 	@Id
 	@Column(name = "ip", nullable = false)
-	private BigInteger ip;
+	private Long ip;
 
-	public BigInteger getIp() {
+	public Long getIp() {
 		return ip;
 	}
 
-	public void setIp(BigInteger ip) {
+	public void setIp(Long ip) {
 		this.ip = ip;
 	}
